Translate religious abilities and the loss thereof
Functionality loss in that "You can now control the weather." from here
on is announced as "You can control the weather."
The ^ screen used them with "You can" prefix and the messaging used
"You can now". Because these are so similar, making both available for
translation would only increase workload for little, if any, gain
(imagine if the two versions go out of sync), considering the other
options this change is probably the best way to proceed.