"Clean up" ability description code
This actually makes it worse.
Checking for a very specific string is not helpful when that string gets
translated before, however. So just always do the magic chopping.
When the HUD is sufficiently internationalized, the "(%s)" part can also
be made available for translations. Right now that'd not make much sense.