Don't extract strings from make_stringf.
It's a bit low level and it's better to mark strings explicitely with _().
In many cases, avoid constructing the string with several static parts.

A couple of minor changes to the english text:
in ugly_thing_mutate(), change "basks in the mutagenic energy from you" to
"basks in your mutagenic energy"
in describe_contamination(), change the description for level 1 from
"very lightly contaminated" to "lightly contaminated" (was duplicated in a
        never reached code path).
Fix inverted plural logic in MiscastEffect::_transmutation